requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
chromadb==0.3.23
click==8.2.2
clickhouse-connect==0.8.18
distro==1.9.0
duckdb==1.3.2
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
hnswlib==0.8.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.10
lz4==4.4.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
posthog==6.3.3
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
zstandard==0.23.0
